initial Puck.World implementation

+++ src/Puck.Launcher/LauncherOptions.cs
@@ -6,8 +6,8 @@ public sealed class LauncherOptions {
     /// <summary>Gets the wall-clock duration after which the terminal auto-exits (for headless or scripted
     /// runs), or <see langword="null"/> to run until closed.</summary>
     public TimeSpan? ExitAfter { get; init; }
-    /// <summary>Gets the target render rate in Hz, or <see langword="null"/> to uncap the framerate (e.g. for VRR). Defaults to 60.</summary>
-    public uint? TargetRenderRate { get; init; } = 60U;
+    /// <summary>Gets the target render rate in Hz, or <see langword="null"/> for automatic display-aware pacing. Defaults to 60.</summary>
+    public double? TargetRenderRate { get; init; } = 60.0;
     /// <summary>Gets whether the genlock phase-aligner is enabled (the former <c>PUCK_GENLOCK</c> toggle, whose
     /// <c>=0</c> value disabled it). Defaults to enabled.</summary>
     public bool GenlockEnabled { get; init; } = true;
